requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
streamlit>=1.31.0
httpx>=0.25.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
langgraph>=0.1.0
starlette>=0.27.0
PyJWT>=2.7.0
jwcrypto>=1.1.0
sse-starlette>=0.7.2
psycopg2-binary>=2.9.6
httpx-sse>=0.2.0